Fix undefined EntityId in BufferingTermLookup
BufferingTermLookup needs the Term objects to contain
the entity id. To do that, we need to fetch both entity id
(numeric) and the entity type.

This patch adds those fields to the selection, and since
this is now identical to getTermsOfEntities, then the code
is changed to have getTermOfEntity to delegate to the
getTermsOfEntities method and fixed the overly long lines.

Also added tests for TermSqlIndex::getEntityTerms.

Bug: T86522
Change-Id: I3092fc5f1eaba4beb9d5db1aa8c1c1aa6ccbb6cc