Update patch set 1
Patch Set 1:

Looks good, except for the name of EntityIdFormatterBase: XyzBase implies an abstract base class (or is that just me?)

I suggest BasicEntityIdFormatter or PlainEntityIdFormatter.

Also, if the interface doesn't support EntityIdValue but only EntityId, EntityIdValueFormatter needs to be implemented separately. A single implementation that delegates to an EntityIdFormatter should be sufficient.

Patch-set: 1
Label: Code-Review=0