Add smoke test in LinkBeginHookHandler
to help catch the problem as occurred in T86522.

BufferingTermLookupTest uses a mock TermIndex,
so the issue was not caught there, and because it
is a lib/store thing, it's not as nice or easy to put
an integration test there with the TermIndex obtained
(e.g.) from WikibaseRepo or WikibaseClient factory.

Change-Id: I811bec57f8909db38d4728f4c2c6f07fe51ff094