Fix change handling for connected pages
if a change is not a site link usage but the page
is connected via a site link, then we still
need to consider it for usage.

we need to do a site link lookup and check if the
page is connected.

(part 2 of bug fix)

Bug: T86765
Change-Id: I445d77bb37cee7ffbd629e4b6ed45da634c3eb6f
(cherry picked from commit ffbb3137ef2b7fb0b1a31f6c283b3316407f4cf3)