Update patch set 2
Patch Set 2:

I think this change would only mask the underlying issue, which (in my opinion) is that the 'sitelinkgroupviewafterstartediting sitelinkgroupviewchange' listener in sitelinkgroupview-sitelinkview does not have all the information it needs to correctly determine if the remove toolbar should be disabled or enabled. The disabling done when starting to save is of a higher rank than the one the 'sitelinkgroupviewafterstartediting sitelinkgroupviewchange' listener manages.

Right now I can't think of a better solution than to have an array of disable reasons in the toolbar and only enabling it when none of the reasons is present. That seems quite complicated to me, and I don't know if I would actually want it. Maybe the ›masking‹ done here is conceptionally worse, but code-wise better.

Patch-set: 2
Reviewer: Adrian Heine <1366@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0