Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

There is not a single "style changes only" file in this patch.

Your "please extract this" comment is about code that was already there. I can split it, sure. But 1st, this is not what this patch is about. And 2nd, I already made this code better. Why do I get a -1 for not making it good enough?

Patch-set: 4