Update patch set 5
Patch Set 5: Code-Review+1

Daniel, I don't have a clue how to do that. These tests are exactly the same as the tests you guys implemented yourself in SpecialItemsWithoutSitelinksTest.php and that was accepted. Can you accept this one and we make a follow up bug/change to improve tests?

Patch-set: 5
Label: Code-Review=+1