Update patch set 2
Patch Set 2: Code-Review+1

You don't have to like it. That's no reason to block this patch, not even with a -1 and clearly not with a -2 (merging it will not break anything, which is almost the only reason to -2 a patch).

This patch replaces a constant with a constant. That's all.

There are already almost 300 instances of the 'property' string constant in our code base. I really don't see a good reason why the few additional places I'm introducing (with quite some benefits, as I extensively explained above) aren't acceptable but the 300 other places are.

The fact that the PHP constant Property::ENTITY_TYPE requires the Property class to be included is broken by design. If some code does have access to the Property class it can typecheck against this class and does not need a string comparison in the first place. The other way around, if code does not have access to a specific Entity class it neither can't do an instanceof check nor a string comparison against these class constants. The only "correct" way is to use a string constant, as this patch does.

Can somebody else have a look at all the arguments given, please?

Patch-set: 2