Update patch set 2
Patch Set 2: Code-Review-2

I agree with Jan: constants are better (better to document, less error prone) than literals. The logical dependency is already there: something defines which values are legal.

As to making it configurable: that does not mean we can or should start using literals. If it's configurable, we should not be using constants or literals, but registries. Code that knows about a speific kind of entity would know and use the constant.

-2 because it's not "please improve", it's simply a bad idea.

Patch-set: 2