Update patch set 1
Patch Set 1:

Not sure about this one...

I would use an instanceof before accessing methods specific to the subclass.

I would check the type constant if filtering by item type - code that does this does not need to know or care about the fact that there are subsclasses of EntityId. It just filters by entity type, which it can get from the EntityId.

It's not horrible to use instanceof for this, but feels conceptually awkward.

Patch-set: 1