Update patch set 5
Patch Set 5:

I'm really scared about how bad our tests for special pages are. The fact that some broken tests were accepted in the past doesn't mean we should accept them now, especialyl since special pages can now use proper CI. 

I understand that it's annoying to be blocked on this, and it's not that easy to come up with a way to test this properly. The SpecialPage base classes make this rather hard. 

I will try to come up with a follow-up that provides better tests. Once that follow-up exists, this patch can be merged. Just filing a ticket isn't good enough I'm afraid, it'll just drop to the bottom of the prio list.

Patch-set: 5