Introduce WikibaseRepo::getTemplateFactory
This is split from Id47127f and will probably make (most of) it
obsolete.

The only unrelated changes in this patch are the added "public" keyword
in the TemplateRegistry constructor (which is and must be public anyway)
and the removal of some unused "use" clauses.

Originally I wanted to introduce TemplateFactory::getDefaultInstance. I'm
following Daniel's request now because I think it's better to have all
global states in one place (the global factory).

I also want to merge the two classes (Templatefactory and TemplateRegistry)
into one class, if possible, but will try to do that in a later patch.

Change-Id: I6876ff253d2ac027b0bf695f1581513c703a7cf1