Drop @since tags from private stuff
I introduced some of these in the previous patch Ide07efc and found
a lot more with a regex. We already did the same in other components
and patches. The information since when a private method or property
is avaliable is not helpfull, speciecially since we barely use
version numbers (still 0.5, oh well).

Change-Id: Ib24e676bcf014639655fe8564783510a6166c44f