Update patch set 6
Patch Set 6:

Made the job delay depend on the max. replag now.

In the future we might want to add a setting to make sure that for example deletions will only be propagated after a minute or so, but this should be good enough to fix the reported bug.

Patch-set: 6