Check for empty arrays in UsageUpdater
(follow-up to I8d52312)

if the list of unused / removed entities is empty,
but added is not empty then the code should still
skip the unsubscribe call, which lead (until I8d52312)
to a database error if using SqlSubscriptionManager.
There still should be a check there, but we can avoid a
bunch of processing by having check here also.

Bug: T88329
Change-Id: Ia2a47e03b8bd10c923a0a911d24541de03aaea8d