Add version constant for EntityContent for parser options
This constant can be bumped for breaking changes in the
ParserOutput of EntityContent. (e.g. changes in EntityView)

This will allow cache invalidation in a more sane way,
in sync with breaking changes in code vs. bumping
$wgCacheEpoch.

Also changed the code to ensure we always have ParserOptions,
consistent with what AbstractContent does in core. Thus,
removed other code for handling null options.

Bug: T86694
Change-Id: I64497f74817e29e9bf4ead73e8f03ad6db5e6c87