Fix -> vs. > typo in DispatchStatsTest
I found the typo because PHPStorm complained about undeclared constants.
Looks like a lot more is broken here.

Please note that I don't fully understand what this test is supposed to
test! I tried to change it so it does succeed on my machine. Let's see
if it succeeds on Jenkins.

Change-Id: I11522d84b88811bdf3fe6aa8d315d18eb96598e9