Update patch set 2
Patch Set 2:

Reverted a few changes resulting in the QUnit tests not passing completely. The mechanism of initializing TemplatedWidgets needs to be improved. Calling the parent's _create should be the last action as that triggers the "create" event which should not be triggered before the widget is initialized completely. This, however, prevents dynamic creation (not on pre-rendered static DOM) of widgets that use TemplatedWidget short-cut selectors during initialization (thus, QUnit tests fail). As the other adjustments of this change set are of high importance, I opted for reverting altering of *view creation routine. Fixing TemplatedWidget creation routine is to be done in another change set.

Patch-set: 2