Update patch set 2
Patch Set 2:

To add more context here: EntityContent::getParserOutput is only being called in case of a cache miss in the first place. So if we add to the parser cache your amended key will be used, but if we try to read from it, it wont, thus this completely disables caching.

Patch-set: 2