Fix type hints in ConnectionManager
This is a re-submission of Ia6f1e10.

Type hinting against an empty (!) interface may have it's use here and
there, but I'm pretty sure it's wrong here. You can't expect an empty
interface to have certain methods. Note that all called methods expect
and return DatabaseBase objects, not IDatabase objects.

This change gets rid of quite some "undefined method" warnings in
related code.

I tried a lot of solutions and this is the best I can come up with. Note
that the abstract base class that declares and (!) implements the methods
declares them as final. This sucks hard because it's impossible to mock
these methods. This leaves me with no solution: The interface is empty
and the next class in the hirarchy is untestable. The only correct
solution is to either kill all "final" keywords from the class or declare
the methods in the interface. Both is a major core patch. I will not
waste my and our time for such a core patch.

This patch also removes two wrong @return tags and some settings in the
test that are the default anyway..

Change-Id: I17f3a445a900a14e1561f70de357636c02122006