Update patch set 1
Patch Set 1:

@jan: This does not remove an optimization, it allows us to continue using the optimization confidently, without introducing inconsistencies and race conditions.

The new behavior is needed to address the issue I outlined in my comment on T88357#1024777: we have an UsageLookup (read) and a UsageTracker (write) service. UsageLookup uses getReadConnection(), so it would usually operate on a slave DB. But if we have just modified the usage table using the UsageTracker, the UsageLookup needs to operate on the master DB, to get consistent results. In this concrete case, we may have removed all usages of an entity by calling trackUsedEntities(), but getUnusedEntities() would still report them as used if it accesses a slave DB.

So we need to somehow make the UsageLookup use the master DB "some times". Namely, always after we modified the data, because we then *know* the slaves are not up to date (unless we waited for them to catch up). 

We could do this by having a second instance of the UsageLookup service which operates on the master DB. Or we could add a parameter to all lookup methods. Both are quite inconvenient, and force us to explicitly pass around state that is really bound to the DB connection. Managing that "dirty" state in the connection manager makes sense, since it "knows" whether a write operation has been performed before.

Patch-set: 1