Update patch set 1
Patch Set 1:

We could also let users of the ConnectionManager remember for which things queries have to go to master. That would possibly reduce the reads to master.

A further possible optimization would be to even cache things in the users of ConnectionManager, so that most just-written changes could directly be read from cache instead of from any db. In such a setup, only reverse queries (for example `entities used by page a`) after writes (for example `entity b uses pages (a, c, d)`) would need to go to master (if deletions are explicit).

That's what jan and I talked about an hour ago. It all depends on the specifics of our queries and the preciousness of master.

Patch-set: 1