Update patch set 1
Patch Set 1:

@Adrian: look at UsageUpdater, that's the concrete use case.

The ConenctionManager doesn't know the queries that get executed, so it can't decide which should go where. But even the service that does know the queries, in our case the UsageLookup, cannot know when to use the master:

The query for unused entities can be run against a slave DB just fine, and should be, per default. The only exception is the when we have just updated the master db. In that case, all subsequent queries need to be executed against master, to ensure consistency (without having to wait for the slaves to catch up).

In the current use case, there are only one or two subsequent queries going to the master DB. Once ConnectionManager is used more widely, we need to keep an eye on this - basically, things that impact each other (access the same tables) should share a ConnectionManager, things that do not should use separate connection managers - so queries against one table would not be affected by updates against another table.

Patch-set: 1