Update patch set 1
Patch Set 1:

This is about the granularity of marking DB_SLAVE as dirty and using DB_MASTER. You suggested to use tables, and we just noted that it's possible to be better than that (by remembering for which entities and pages data was written), thereby reducing the number of queries against master (only those referencing an entity or page for which data was written). It would increase complexity, so we asked how bad it is to query master. It's a trade-off.

Patch-set: 1