Update patch set 1
Patch Set 1:

I see what you mean, but I don't think that added complexity is remotely worth it, at least not for the current use case. We are talking about a single table (wbc_entity_usage) anyway.

Once ConnectionManager is used for more complex cases, this is worth investigating, but just marking tables as "dirty" is not going to cut it, I think. You *can* have inconsistencies between tables, even if the one you are reading from was not touched directly.

Patch-set: 1