Update patch set 1
Patch Set 1:

Just to be clear: You suggested to use tables: »Once ConnectionManager is used more widely, we need to keep an eye on this - basically, things that impact each other (access the same tables) should share a ConnectionManager«

I'd probably agree that the added complexity is not worth it right now. I'm not sure that the ConnectionManager is the right one to make _this_ decision, since it's definitely not the right one to make more fine-grained decisions.

Patch-set: 1