Update patch set 1
Patch Set 1:

Adrian: I see what you mean, but the alternatives (an additional service instance, along with logic to decide when to use it, or extra parameters everywhere) seemed exceedingly bad.

And while I agree that the ConenctionManager cannot *make* the more fine grained decisions, it does seem like a good place to maintain the *state* that results from these decisions. 

In general, I think that "after writing to master, always read from master" is a good rule of thumb; it's certainly true per table, and sometimes for groups of tables. And as long as one ConnectionManager instance is used per such logical group (which is trivially the case here), the performance impact should be minimal.

Patch-set: 1
Label: Verified=0