Add fail-safe check to Utils::getLanguageCodes
Core's Language::fetchLanguageNames() probably never returns an empty
array, but who knows? Very odd, very hard to track things happen in
Wikibase if this is ever the case. So better add this check and throw
an exception as early as possible.

This is the result of a discussion I had with Daniel: What to do if a
list of "allowed codes" is empty?
1. Empty array means the same as null: No filter, allow everything
   (my favorite, but not an option for Daniel).
2. Allow nothing (Daniel's favorite, but not an option in my opinion).
3. Neither nor (this patch).

This is split from Ic506adc.

Change-Id: I3507a4604960588faa3932fe43d9c2509ae27650