Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Sorry for jumping in, the number of people involved here is probably high enough. But I see a major problem with this patch:

Code that really, really need to do a read operation after doing a write operation should simply re-use the write connection. Simple as that. On the other hand, code that does not care should not be forced to read from master if not necessary.

The behavior you are introducing here is nothing I expect from something called "ConnectionManager". If I ask a ConnectionManager for a read-only connection it should not make any assumptions about what I'm going to do with and expect from this read-only connection.

What you are introducing here is probably a TransactionManager. I'm not sure. I would expect something like "getTransactionManager" and on top of that the two "getRead" and "getWriteConnection" with the behavior you are introducing here. Please rename the class accordingly or introduce a new one with a name that makes it clear that independent pieces of code should not reuse the same manager.

Patch-set: 2
Label: Code-Review=-1