Update patch set 2
Patch Set 2:

@Thiemo: your argument is essentially the same that came from Adrian and Jan, which we resolved yesterday. Are you insisting that the logic is inherently bad? Or do you just want a better name? From your comment, it's not clear to me whether renaming would be sufficient to address your concern.

As to renaming... TransactionManager is a bit odd considering the interface, but I can't think of anything better. ConnectionManager is actually closer to the facts. But I'm fine with renaming this.

By the way: no assumptions are made about what the caller does with or expects from the read connection. The assumption that *is* made is that after modifying the database, read operations should not be performed against stale slaves. I think this logic is sound and it would be perfectly valid to implement it globally - the only downside to that is the performance cost (which can be addressed in the manner described).

Patch-set: 2