Update patch set 2
Patch Set 2:

> your argument is essentially the same

Yes, I'm aware of that.

> which we resolved yesterday.

Resolved how?

> do you just want a better name?

Yes. Please rename the class to ConsistentReadConnectionManager and I will merge this. ConnectionManager can be an interface, but we don't need this at the moment and can easily introduce it later, if needed.

> no assumptions are made about what the caller [...] expects from the read connection.
> The assumption that *is* made [...] read operations should not be performed against stale slaves.

That's the same. How do you know all callers expect that from a read connection? An assumption is an assumption, and it's currently only communicated through comments.

Patch-set: 2