Update patch set 2
Patch Set 2:

@Thiemo: ok, I'll rename.

Re the assumption: when asking for a read connection, I can expect that I can run select queries against it. That's it. Any expectation that this would not be a master connection would be misguided. There is no place in core where this is guaranteed.

Patch-set: 2
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0