Localize calendar names in TimeDetailsFormatter
The messages are defined in ValueView. I'm not sure if it's a good
idea to reuse them. This is not a problem now but could become one
when we try to split or move this code into components.

This patch drops the options thingy because this is not used anyway
and I think it is not usefull.

Change-Id: Icbd60f316f55171e06c6c68ec3d2e25f7e605a5e