referenceview: Removed internal listview caching
Caching listview and its listItemAdapter is prone to errors as the listview might get regenerated
and the cached reference needs to be updated on every occasion. Even more, the concept of caching
the listItemAdapater as option is conceptually wrong.
listview instance and listItemAdapter should always be retrieved from the $listview node.

Change-Id: I843a44a470f1da133c522230e4791b21e4f898ed