Update patch set 2
Patch Set 2: Code-Review-1

Looks good in general, but I'm not a fan of how you have both "monolingualTextLanguages" and "ContentLanguages".

Either it's an abstract concept (which it actually is as of this path, those are the same languages we use for Terms, so we should use that class also there) or it's a concept which only applies to MonolingualText (which could be valid if those are going to be handled differently from eg. Terms).

Patch-set: 2