Update patch set 4
Patch Set 4:

Thiemo was right, I cheated a bit to make the sorting simpler, amended it now to always sort on group id, then on site id.
But I chose to not go the route of amending the sitelinklookup for various reasons. If you want to do that, please go ahead, but it will be way to much work for this.

Patch-set: 4