Drop unused code from deprecated Wikibase\Term class
This class should not be used any more. This patch drops methods
(or makes fields and methods private) if they are not used any more
outside of tests.

Originally I wanted to drop the static compare method but it's still
used. Instead I simplify it a lot to not do a full comparison but a
much simpler equals check instead.

Change-Id: I861ad8571f77814814fb63d7b2f06e6f4ac04df1