Update patch set 2
Patch Set 2:

Agree to Adrian, -2 on the process. Two people decided within hours and nobody had a chance to keep up. Hoo, I already told you that half a year ago, and the same applies to Jeroen: If a change is not super-trivial, please do *not* insta-merge. We are a team and supposed to work in a team. Interface changes need discussion and proper review.

About the interface discussed here: I tried to understand what the problem with the interface was and still don't get it. There is a method hasLanguage($code). Why is this ok but a method getName($code, $inLanguage) is not? What the heck? Whats the difference? I really don't get it.

In general, I still (since I started working here) have a problem with the injection hell we are constantly building. Why do we need sooo many services, each with a single method?

Patch-set: 2