Schedule a LinksUpdate job if sitelinks changed
And no longer purge the page.

We handle langlinks using ParserOuput these days
so that this is fine for the UI and it also makes
the API output (action=query&prop=langlinks) work.

I tested that unlinking and initially linking a client
page does still trigger all relevant updates (which is
more than LinksUpdate).

Bug: T89903
Change-Id: I6b8b7bce1f2843af0fa43db24af9c71a2c3e875c