Set Change timestamp based on RC timestamp.
This was intended to set the correct timestamp for undeletion events.
However, RC entries for undeletion currently do not have a revision ID
associated, making it impossible to find the corresponding entry in
the wb_changes table.

So, this change is provided for completeness, but does not do much at
the moment.

Change-Id: If8190516cb998b58ca7b762b3293e31d6102be9a