Make h1 header css in entitytermsview more full width
These changes appear to do what's desired in T90426,
making the header and bottom-border full width of
wikibase-entityview-main while leaving margin for
the edit and save|cancel buttons.

Ended up using margin for the h1 element, since
the edit toolbar is sticky and when scrolling can
result in ugly overlap with the h1 bottom border.

Bug: T90426
Change-Id: Ie82f3b74334b8db2c6bdcafc350f7505dc46c83d