Make h1 header css in entitytermsview more full width
This gives the header and bottom-border full width
of wikibase-entityview-main.

This causes overlap of the edit toolbar with very
long labels, descriptions and aliases.

That could be fixed by adding a right margin to
the label, aliases and descriptions, or possibly
a sticky edit toolbar with a background and higher
z-index.

Bug: T90426
Change-Id: Ie82f3b74334b8db2c6bdcafc350f7505dc46c83d