Add fail-safe for YYYY-00-00 to DateTimeParser
PHP's internal DateTime/strtotime parser does actual calculations
that can't be turned off. For example, 2015-00-00 becomes 2014-12-30,
which is horrible in our use-case. This patch does not detect all
cases where this happens (2015-12-00 is still parsed wrong, for
example) but at least some pretty relevant ones. The idea is simple:
There can only be one 3+ digit number in a date time string: the year.
Extract the year and check if it is still present in the output.

Bug: T87574
Change-Id: I0054052cd3a8d33a7d11b018ad84f79a7a3a7df9