Rename DateTimeParser to PhpDateTimeParser
This is not just *A* date time parser, it uses *THE* DateTime object
from PHP's core to parse date strings. I want this to be reflected in
the name of the class to make it obvious that it should only be used
as a fallback if no other, more specific parser could be applied.

Change-Id: I4e997a3af824b124c77d3eaf2d7ac08be5e7b412