Fix overlap of edit button and aliases text
This gives the edit toolbar a white background, which
together with the z-index, keeps the edit toolbar clearly
separated and "above" the aliases text.

This is not totally ideal, as it might be nice for
the edit buttons to get the blue background when one
scrolls down to the "other languages" box and that
becomes sticky.  I think trying to do that much in
this patch would overcomplicate it and this solution
is sufficient for now.

Bug: T89956
Change-Id: Ideafcd1d6b20e180cd3cec1048d7071ee085a691