Update patch set 18
Patch Set 18:

@Stas: oh, if you not don yet, fine :)

By the way, I didn't mean to say that all combinations must be tested. But there should be *some* tests for the other modes.

Anyway, if the patch is work in progress, best give yourself a -1 with a comment like "need more test". Then it's clear that reviews would be helpful, but you don't consider the patch read yet.

Patch-set: 18