More consistent layout w/ collapsed "other languages" and no statements
Increased (and changed to use "margin") for description and alias text,
to accommodate more lengthy words like "[speichern | abbrechen]"
in the edit toolbar.

Set width of .wikibase-entityview-main to 100%, so that
the header is more consistent when collapsing "other languages"
and expanding.

This also keeps some stuff from overflowing
which keeps some minimum width for now, even if it means
overflow. If the statements become too narrow, then
we get strange wrapping of words, like:

Q
1
1
8

especially in qualifiers.

Since we are planning to do major redesign of the statements
section, fixing these issues can be part of that and done
later.

Will make a separate patch to address overflow of the sitelinks
section.

Tested this with Firefox and Chrome, with browser window at
various sizes.

Bug: T90395
Change-Id: I5c7e037a6f99a8b704b89a22151756d572bede1a