Update patch set 6
Patch Set 6:

Removing '?>' broke the Jenkins build since it's not superfluous. It would be superfluous if it was creating a new PHP file or if it was the last thing every to run, but it's amending a pre-existing LocalSettings.php. Earlier in the same script here it is also appending '<?php' to LocalSettings.php. That would break if it didn't already end in '?>'. As such it should maintain that for the next writer. For legacy reasons (and to allow templates to be valid php files), all writers to LocalSettings.php begin and end '?>'. @jzerebecki is looking into this now :)

Patch-set: 6
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0