Use EntityIdFormatter in MessageParameterFormatter
MessageParameterFormatter currently does [[TITLE|ENTITYID]]. This change makes
it use a EntityIdLinkFormatter, which used to do [[ENTITYID]]. This change also
makes it do [[TITLE|ENTITYID]] instead. EntityIdLinkFormatter is currently only
used in the SummaryFormatter.

Put together, the only change in behaviour this change introduces is the
trimming of the namespace in links to entities in summaries.

Change-Id: Ia5c1a4d19c4c40ce668840700c032ce4f1302f9b