Automatically split parser cache.
Calling getters on ParserOptions can reigged to automatically split
the parser cache according to the respective options. This is done
by registering the ParserOutput object as a watcher to the ParserOptions.

Note that the automatic cache works for some options (e.g. userlang, thumbsize)
but not for others (e.g. editsection).

Change-Id: Ide34f04eff772c098194f63241b6514bdd403e86