Update patch set 2
Patch Set 2: Code-Review-1

I find the new code a lot harder to understand than the old one. Expressing something in the most compact way is not the same as simplifying it.

The old code wasn't pretty, but the new code is incomprehensible... Maybe that can be fixed with appropriate comments. Or by not trying to cover *everything* with a single regex...

Patch-set: 2
Label: Code-Review=-1