Update patch set 2
Patch Set 2:

There is a comment. Come on. There is nothing fancy in this regex. Maybe you are missing a test? I don't see an obvious way forward here. How does splitting a regex at the '|' make it easier to understand if it's still the exact same regex, except that it's split?

Patch-set: 2
Label: Verified=0