Maintenance script for updating subscriptions.
This maintenance script updates a client wiki's subscriptions stored in the wb_change_subscription table on the repo, based on the information in the
client's wbc_entity_usage table.

This script is intended to be run some time after the populateEntityusage script
when enabling usage tracking on a client wiki.

Bug: T90756
Change-Id: I073dd2b6ee9ae313d88545c658ca9729a8edb049