Update patch set 1
Patch Set 1:

I do not think that the listview should be created in non-edit mode when there are no qualifiers as that may result in styling issues as well as some performance decrease since creating the widget causes events to be triggered for no proper reason. The particular if-clause is changed in https://gerrit.wikimedia.org/r/#/c/190799/ as well. Not sure whether that would turn out to be a better alternative.

Patch-set: 1
Reviewer: Henning Snater <167@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0