Update patch set 2
Patch Set 2:

(1 comment)

Will fix the wrong import.

There is a test that checks that the correct flags are present in the ParserOutput. The meaning of this is however not well specified or tested in core. Technically, we are testing that ParserOutput::getUsedOptions() reports the correct options "which have been taken into account to produce this output". That's the contract we rely on. We expect core to split the cache based on this information, be we do not test that. That should be tested by core (but isn't, of course).

Of we want to test full integration with the parser cache, I suggest a browser test. Want to file a ticket for that?

Patch-set: 2
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0