Update patch set 2
Patch Set 2:

If you give the wrong type of parameter when calling a method, and this would cause the method to behave strangely, then the method should throw an exception, not try to do "something nice".

This is not about handling invalid data structures form the database. This is about violating the contract of the method.

Patch-set: 2