Update patch set 2
Patch Set 2:

I would have to copy-paste the same code 3 times. I would need to introduce at least 3 new variables and more stuff to make this work.

Think of it like it is a range check, not a type check. It's the responsibility of the method to react in a reasonable way if a value is out of range. You do not copy-paste the range check all over the place.

Patch-set: 2