Update patch set 2
Patch Set 2:

Wait, I think I understand better now.
This method is used for two distinct purposes: outputting the precision, *and* outputting the value with a precision. 

For outputting the precision, returning a "bad" precision as-is seems fine. 

But the two use cases should just be two separate function. Why mash them together in one function? It leads to a strange soup of conditionals and optional things.

Patch-set: 2