Update patch set 2
Patch Set 2:

I'm afraid my current approach of creating a long chain of very small patches is prone to bike sheding. :-(

There is nothing "mashed".

What the precision "year" really means is "1 year". That's what the function displays.

The before and after values are to be measured in the precision. An after value of "2" means "2 years". Same message. Same code. Why have to functions that are 99% identical?

Patch-set: 2