Update patch set 1
Patch Set 1:

> it should take me to SetLabelDescriptionAliases/Q139/en.

This is not possible without JavaScript or at least a redirect, which will make it harder to use the browsers back button. I'm not sure if this is a good idea. It would be kind of nice, but is out of the scope of this patch. It's an unrelated feature request that needs a separate ticket.

> instead of "guess" (en) a language [...] language needs to be selectable.

Language is selectable. You can type whatever you want. It just defaults to the current UI language, which is most probably what you want.

> why are the input boxes so small?

Are they? A size of 50 is already pretty big. Was defined to be like that a long time ago (January 2013).

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0