Make special page input fields wide and responsive
The "size" attributes are almost pointless if there is a CSS width
defined, which is the case. Removing it makes the code simpler and
easier to maintain.

Browsers default is 40, by the way.

I'm also making the fields much wider (and responsive, which wasn't
the case before), which is relevant for label, description and alias
fields.

Change-Id: Idd798402dac7df27ecf4e7dd56ff4f5b3ad35af4