Drop pointless time formatter delegation
The so called "TimeFormatter" lives in the Time component and doesn't
do much: It formats the TimeValue with the given formatter or uses the
unformatted timestamp. And it appends "(Gregorian)", which is just
wrong and must be fixed anyway. But even if "TimeFormatter" is fixed,
it should not be used here. It's a plain-text formatter, but this code
needs a HTML formatter.

I'm fixing the problem in two steps:

1. This patch removes the appended "(Gregorian)". This fixes an actual
bug I'm trying to fix for a month now. Nothing is lost because the
detail formatter displays the calendar model anyway.

2. The next patch uses a proper HTML formatter.

Bug: T92265
Change-Id: I7e467109532997593b3f30f82899ab781c038215