Make "in more languages" box toggle faaast
Having an animation is very nice in this case, but please don't waste
users time and make this fast. This is not a complicated animation that
the user needs to understand. It just toggles an element on and off,
that's it. jQuerys default animation speed is waaay to slow for almost
everything.

Change-Id: I73ac468eb5395cfe725a106510fd647f9a0f38c0