Add white background to overlapping edit buttons
This is an alternative proposal to Ideafcd1. This solves the problem
of the edit buttons becoming unreadable. It's not a "perfect" solution,
especially because it does not fix the problem of the text *under* the
edit buttons becoming unreadable. This is on purpose. I do not want to
add a padding somwhere that wastes space and is only useful in certain
edge cases.

Note that this issue will probably be solved with the new UI.

Bug: T89956
Change-Id: I9b98115efe337218404b35358e80ab67df99cbe8