Don't call JobQueue::delayedJobsEnabled
JobQueue::delayedJobsEnabled was removed, so don't call it.
We no always set the jobReleaseTimestamp param on the job.
JobQueue implementations are free to use or ignore it.

NOTE: this fixes a fatal error caused by calling a non-existing method.
The core change that caused this is I3a1a78bb48b.

Change-Id: I840d3cc81c3d6f25e019880166dddcad23612f4f