Disable JSCS requireSpaceAfterLineComment
This is exactly the type of completely over-the-top, completely irrelevant
rule that makes me go nuts. Blocking (!) a patch with a -2 (!) because of
semantically irrelevant whitespace (!) in semantically irrelevant single-line
comments (!). What on earth? We have several hundrets (!) of single-line
comments with no space in our code base. Most of them in PHP. We may disagree
if such comments are good-practice (I ommit the space if the comment is
behind some code and add it if the comment is on it's own line). But this is
not an error and not a reason to waste peoples time.

Change-Id: I239a8e9089887ca73dcd1f2bc1ce0d1ab112ffc9