Include WikibaseView directly in repo entry point
the WikidataBuild avoids using the 'magic' composer
entry point, so does not load view automatically when
done that way.

This changes the loading to be the way WikibaseLib is
loaded. (not ideal, but works and more reliable)

When WikibaseView is completely split, then it would
be listed in composer.json and loaded that way instead.

Change-Id: I34db2a41597d35be3c9cfe56298c5648c8b7d4ce