Update patch set 2
Patch Set 2: Code-Review+1

@Adrian: I'm afraid you missed my point. Please let's talk in person.

I don't have a problem with a style guide that says "please use spaces in comments".

I do have a problem with a tool that hits me with a hammer because of a goddamn irrelevant space in a goddamn irrelevant comment. The "Verified: -1 Fails" rating on Gerrit is for failing tests. It's an alarm that a patch is horrible and will break production.

But a goddamn space in a goddamn comment is not the same as a broken test.

A missed space is more like a FIXME. It doesn't do anything. It's irrelevant in production. It can be merged and fixed later. Setting up a rule that actively blocks patches because they contain a FIXME does not make any sense. Same is true for whitespace in comments.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1