Update patch set 2
Patch Set 2:

Unless this rule was indadvertedly introduced due to auto-updating dependencies, I assume 'master' is currently passing this.

Disabling it until the code base passes it makes perfect sense. However it is established practice in many open source environments (including Wikimedia, jQuery, Bootstrap, Angular, Gruntjs, and maybe others) that coding style is required to pass before merging the commit. For all the provided reasons and more, it is decided that code must look coherent and consistent before submission to the repository. The only exception to this is security releases and live hacks applied directory to the cluster that bypass Gerrit. That is the reality.

Patch-set: 2