Update patch set 2
Patch Set 2:

Also note that many editors support plugins for automatic and live correction of coding style as you type (using auto-discovery for the nearest jscsrc file). And of course one can run 'npm test' or 'composer test' pre-commit locally to reduce the overhead of waiting for Jenkins to come back with a style error. That is understandably frustrating, but can be improved in the local workflow.

Patch-set: 2